Free UART Zero
This change removes the special-casing of UART0.

Instead, another slot is added to the list supported by uos.dupterm;
UART0 is added to slot 1 when booting if there are no slots in use.

This replaces our half-baked UART0 IRQ handler with a timer that checks
for new console characters 20 times per second.