extmod/uasyncio/stream: Add a 'Stream.awrite' method.
The usual write+drain combo is suboptimal and causes excessive data
copying, esp. when you write a memory buffer.

Fixing this within the existing code is nontrivial, thus this commit
adds a new 'Stream.awrite' method that can (indeed should, if CPython
compatibility is not an issue) be used instead.